Add Manual Entry page
The manual entry page is built with some intuition, to
automatically assign the most likely tag based on the
text entered for the receipt item.

To support this intuition, a new table, CommonTag, had to be created.
This table contains pairs of 'name' and 'tag'. Each name is unique,
and the tags are not. The idea is that given a name, you can find the
tag that's associated with it.

These pairings are created by looking at the last two weeks of receipts
in the database, and figuring out which tag is most associated with
a name through SystemItems. This is updated daily as a job.